+++ lib/signals/base_signal.dart
@@ -20,6 +20,9 @@ abstract class BaseSignal<T> {
 
   BaseSignal.deferred();
 
+  /// True if the signal has a value assigned.
+  bool get hasValue => _isInitialized;
+
   T get value => _value;
 
   @protected
@@ -33,9 +36,8 @@ abstract class BaseSignal<T> {
 
   Stream<T> get stream => controller.stream;
 
-  Stream<T> get streamTriggerImmediatly => _isInitialized
-      ? controller.stream.startWith(_value)
-      : controller.stream;
+  Stream<T> get streamTriggerImmediatly =>
+      _isInitialized ? controller.stream.startWith(_value) : controller.stream;
 
   /// Returns true if this signal has already been disposed.
   bool get isDisposed => controller.isClosed;