+++ lib/signals/signal.dart
@@ -49,7 +49,7 @@ class Signal<T> extends BaseSignal<T> {
   @protected
   void emit(T newValue) {
     if (isDisposed) return; // Node was disposed before the async op completed
-    if (unsafeValue == newValue) return; // Small optional optimization
+    if (hasValue && unsafeValue == newValue) return; // Small optional optimization
     unsafeValue = newValue;
     controller.add(newValue);
   }